Moved nested modules into just /models/module.py format. Updated Tests to reflect change in paths.
