--- app/models/User/__init__.py
@@ -1,1 +1,0 @@
-import app.models.User.user
