+++ app/models/__init__.py
@@ -1,13 +1,13 @@
-from app.models.Note.note import Note
-from app.models.Message.message import Message
-from app.models.User.user import User
-from app.models.Analytics.analytics import Analytics
-from app.models.Benefits.benefits import Benefits
-from app.models.KeyManagement.key_management import KeyManagement
-from app.models.PaidTimeOff.paid_time_off import PaidTimeOff
-from app.models.Pay.pay import Pay
-from app.models.Performance.performance import Performance
-from app.models.Retirement.retirement import Retirement
-from app.models.Schedule.schedule import Schedule
-from app.models.WorkInfo.work_info import WorkInfo
+from app.models.note import Note
+from app.models.message import Message
+from app.models.user import User
+from app.models.analytics import Analytics
+from app.models.benefits import Benefits
+from app.models.key_management import KeyManagement
+from app.models.paid_time_off import PaidTimeOff
+from app.models.pay import Pay
+from app.models.performance import Performance
+from app.models.retirement import Retirement
+from app.models.schedule import Schedule
+from app.models.work_info import WorkInfo
 import app.signals
