+++ app/models/analytics.py
