+++ app/tests/models/test_model_note.py
@@ -10,7 +10,7 @@ from app.tests.mixins import Pep8ModelTests
 
 class NoteModelTests(TestCase, Pep8ModelTests):
     def setUp(self):
-        self.path = "app/models/Note/note.py"
+        self.path = "app/models/note.py"
 
     def test_can_create_note(self):
         note = Note(note_name="Hey!", pub_date=timezone.now())
