+++ pygoat/production_settings.py
@@ -1,9 +1,9 @@
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.mysql',
-        'NAME': '',
-        'USER': '',
-        'PASSWORD': '',
+        'NAME': 'DjanGoat',
+        'USER': 'djangoat',
+        'PASSWORD': 'default1!',
         'HOST': '',  # can leave blank (sets to default)
         'PORT': '',  # can leave blank (sets to default)
     }
